remove declarations page
