+++ app/controllers/citizens/check_answers_controller.rb
@@ -4,13 +4,28 @@ module Citizens
       legal_aid_application.check_citizen_answers! unless legal_aid_application.checking_citizen_answers?
     end
 
+    # def continue
+    #   go_forward
+    # end
     def continue
+      record_acceptance
+      legal_aid_application.complete_non_passported_means! unless legal_aid_application.provider_assessing_means?
+      ProviderEmailService.new(legal_aid_application).send_email
+      CitizenCompleteMeansJob.perform_later(legal_aid_application.id)
       go_forward
     end
 
     def reset
       legal_aid_application.reset!
       redirect_to back_path
+    end
+
+    private
+
+    def record_acceptance
+      return if legal_aid_application.declaration_accepted_at?
+
+      legal_aid_application.update!(declaration_accepted_at: Time.now)
     end
   end
 end
