--- app/views/citizens/declarations/show.html.erb
@@ -1,11 +1,0 @@
-<%= page_template page_title: t('.page_heading') do %>
-  <%= render 'shared/applicant_declaration' %>
-
-  <%= next_action_buttons_with_form(
-        url: citizens_declaration_path,
-        method: :patch,
-        show_draft: false,
-        continue_button_text: t('.submit_button')
-      ) %>
-
-<% end %>
